+++ core/data/src/main/kotlin/de/tum/informatics/www1/artemis/native_app/core/data/service/impl/courses/CourseRegistrationServiceImpl.kt
@@ -41,7 +41,7 @@ internal class CourseRegistrationServiceImpl(
     override suspend fun registerInCourse(
         serverUrl: String,
         authToken: String,
-        courseId: Int
+        courseId: Long
     ): NetworkResponse<Unit> {
         return performNetworkCall {
             val user: Account = ktorProvider.ktorClient
