+++ core/model/src/main/kotlin/de/tum/informatics/www1/artemis/native_app/core/model/exercise/FileUploadExercise.kt
@@ -10,7 +10,7 @@ import kotlinx.serialization.Serializable
 @Serializable
 @SerialName("file-upload")
 data class FileUploadExercise(
-    override val id: Int? = null,
+    override val id: Long? = null,
     override val title: String? = null,
     override val shortName: String? = null,
     override val maxPoints: Float? = null,
