+++ core/websocket/src/main/kotlin/de/tum/informatics/www1/artemis/native_app/core/websocket/ParticipationService.kt
@@ -34,7 +34,7 @@ interface ParticipationService {
      */
     fun getLatestPendingSubmissionByParticipationIdFlow(
         participationId: Int,
-        exerciseId: Int,
+        exerciseId: Long,
         personal: Boolean,
         fetchPending: Boolean = true
     ): Flow<ProgrammingSubmissionStateData?>
