+++ feature/course_view/src/main/kotlin/de/tum/informatics/www1/artemis/native_app/feature/course_view/CourseViewModel.kt
@@ -17,7 +17,7 @@ import java.util.Locale
 
 
 internal class CourseViewModel(
-    private val courseId: Int,
+    private val courseId: Long,
     private val serverConfigurationService: ServerConfigurationService,
     private val accountService: AccountService,
     private val courseService: CourseService,
